chrome extension built

+++ js/main.js
@@ -1,51 +1,100 @@
 document.getElementById("myForm").addEventListener('submit', saveBookmark);
+window.addEventListener('load', saveBookmark1);
+
+
+
+function saveBookmark1(e)
+{
+  //alert(document.getElementById("myForm"));
+  //alert("hello")
+  chrome.tabs.getSelected(null,function(tab) { // null defaults to current window
+    var title = tab.title;
+    var url = tab.url;
+    document.getElementById('siteName').value = title;
+    document.getElementById('siteUrl').value = url;
+
+  });
+
+}
+
 
 function saveBookmark(e)
 {
+  // var currentUrl = window.location.href;
+  // alert("vjdnvj");
+  //console.log(currentUrl);
   var siteName = document.getElementById('siteName').value;
   var siteUrl = document.getElementById('siteUrl').value;
+  //alert(siteUrl);
+
 
   if(!validateForm(siteName, siteUrl))
   {
     return false
   }
 
-  if(!siteUrl.includes("http://"))
+  if(!siteUrl.includes("http://") && !siteUrl.includes("https://") && siteUrl.includes("www."))
   {
-    siteUrl = "http://"+siteUrl;
+    siteUrl = "https://"+siteUrl;
   }
-
 
     var bookmark = {
     name:siteName,
     url:siteUrl
   }
 
-  //console.log(bookmark);
-
+  //alert(bookmark)
+  //
+  // //console.log(bookmark);
+  //
   if(localStorage.getItem('bookmarks')===null)
   {
     var bookmarks = [];
     bookmarks.push(bookmark);
     localStorage.setItem('bookmarks', JSON.stringify(bookmarks));
     //localStorage.setItem('bookmarks', bookmarks);
-    console.log(bookmarks);
+    //console.log(bookmarks);
 
   }
   else {
     var bookmarks = JSON.parse(localStorage.getItem('bookmarks'));
 
-    bookmarks.push(bookmark);
-    localStorage.setItem('bookmarks', JSON.stringify(bookmarks));
-    console.log(bookmarks);
+    if(!bookmarkExist(bookmarks, bookmark))
+    {
+      bookmarks.push(bookmark);
+      localStorage.setItem('bookmarks', JSON.stringify(bookmarks));
+    }
+    else {
+      alert("already exist");
+    }
+
+    //console.log(bookmarks);
   }
+  //
+  //
+  //document.getElementById('myForm').reset();
+  // fetchBookmarks();
+  // //e.preventDefault();
 
+  //alert(localStorage.getItem('bookmarks'));
 
-  document.getElementById('myForm').reset();
-  fetchBookmarks();
-  //e.preventDefault();
 
 }
+
+function bookmarkExist(bookmarks, bookmark)
+{
+  for (var i = 0; i < bookmarks.length; i++) {
+    name = bookmarks[i].name;
+    url = bookmarks[i].url;
+
+    if(name==bookmark.name)
+    {
+      return true;
+    }
+  }
+  return false;
+}
+
 
 function validateForm(siteName, siteUrl)
 {
@@ -64,41 +113,4 @@ function validateForm(siteName, siteUrl)
   }
   return true
 
-}
-
-function deleteBookmark(name)
-{
-  var bookmarks = JSON.parse(localStorage.getItem('bookmarks'));
-  if(bookmarks!=null)
-  {
-    for (var i = 0; i < bookmarks.length; i++)
-    {
-      if(bookmarks[i].name==name)
-      {
-        bookmarks.splice(i,1);
-      }
-    }
-    localStorage.setItem('bookmarks', JSON.stringify(bookmarks));
-  }
-  fetchBookmarks();
-}
-
-
-function fetchBookmarks()
-{
-  var bookmarks = JSON.parse(localStorage.getItem('bookmarks'));
-  if(bookmarks!=null)
-  {
-    var results = document.getElementById('bookmarksResults');
-    results.innerHTML = '';
-    for (var i = 0; i < bookmarks.length; i++) {
-      name = bookmarks[i].name;
-      url = bookmarks[i].url;
-      console.log(url);
-
-      results.innerHTML += '<div class="well"><h3>'+name+' <a class="btn btn-default" target="_blank" href="'+url+'">visit</a> '+
-                           ' <a onclick="deleteBookmark(\''+name+'\')" class="btn btn-danger">delete</a></h3></div';
-
-    }
-  }
 }
